Adjusted Title
bump version
Added a Privacy Nutrition domain
Readding the export transaction data
Export Layout for the Transaction Data
Polishing the Trans Detail
Fixed the no send when syncing.
Removed ancient TabBarViewController
Adjust the general purpose framework

+++ brainwallet/Transaction.swift
@@ -23,7 +23,7 @@ class Transaction : Hashable, Identifiable {
 		self.tx = tx
 		self.wallet = wallet
 		self.kvStore = kvStore
-		let fee = wallet.feeForTx(tx) ?? 0
+		let fee = wallet.feeForTx(tx) ?? UInt64(0)
 
 		var outputAddresses = Set<String>()
 		var opsAmount = UInt64(0)
